driveamerican/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-autoslug==1.9.6
django-crispy-forms==1.8.1
django-froala-editor==3.1.0
djangorestframework==3.11.0
pyTelegramBotAPI==3.6.7
pytz==2019.3
requests==2.7.0
six==1.14.0
sqlparse==0.3.0
